requirements.txt
streamlit==1.52.2
Pillow==12.0.0
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.3
python-dotenv==1.2.1
google-genai==1.56.0